+++ game_model/src/main/scala/se/gigurra/wallace/gamemodel/TerrainStorage.scala
@@ -1,13 +1,13 @@
 package se.gigurra.wallace.gamemodel
 
 trait TerrainStoring[T_TerrainStorage] {
-  def width(t: T_TerrainStorage): Int
-  def height(t: T_TerrainStorage): Int
-  def get(t: T_TerrainStorage, iPatch: Int): TerrainPatch
-  def set(t: T_TerrainStorage, iPatch: Int, patch: TerrainPatch): Unit
+  def width(t: T_TerrainStorage): Long
+  def height(t: T_TerrainStorage): Long
+  def get(t: T_TerrainStorage, iPatch: Long): TerrainPatch
+  def set(t: T_TerrainStorage, iPatch: Long, patch: TerrainPatch): Unit
 }
 
 trait TerrainStorageFactory[T_TerrainStorage] {
-  def create(wPath: Int, hPath: Int): T_TerrainStorage
+  def create(wPath: Long, hPath: Long): T_TerrainStorage
 }
