+++ game_model/src/main/scala/se/gigurra/wallace/gamemodel/WorldVector.scala
@@ -4,5 +4,5 @@ import se.gigurra.wallace.gamemodel
 
 object WorldVector {
   import gamemodel.WorldVector
-  def apply(x: Int = 0, y: Int = 0): WorldVector = new WorldVector(x, y)
+  def apply(x: Long = 0, y: Long = 0): WorldVector = new WorldVector(x, y)
 }
